requirements.txt
streamlit
torch
transformers
python-dotenv
openai-whisper
sounddevice
numpy
google-genai
